Rename file and fix rendering issues
