+++ spec/draft/API_specification/inspection.rst
@@ -3,21 +3,36 @@ Inspection
 
     Array API specification for namespace inspection utilities.
 
-A conforming implementation of the array API standard must provide and support the following functions and objects.
+A conforming implementation of the array API standard must provide and support the following functions and associated inspection APIs.
 
 
 Objects in API
 --------------
 
-.. currentmodule:: array_api
+.. currentmodule:: array_api.info
 
 ..
   NOTE: please keep the functions in alphabetical order
 
 .. autosummary::
    :toctree: generated
-   :template: attribute.rst
-   :nosignatures:
+   :template: method.rst
 
    __array_namespace_info__
-   Info
+
+
+Inspection APIs
+---------------
+..
+  NOTE: please keep the methods in alphabetical order
+
+
+.. autosummary::
+   :toctree: generated
+   :template: method.rst
+
+   capabilities
+   default_device
+   default_dtypes
+   devices
+   dtypes
