--- src/array_api_stubs/_draft/inspection.py
@@ -1,139 +1,0 @@
-from __future__ import annotations
-
-__all__ = ["__array_namespace_info__", "Info"]
-
-from ._types import (
-    Optional,
-    device,
-    dtype,
-    DefaultDataTypes,
-    DataTypes,
-    Capabilities,
-)
-
-
-def __array_namespace_info__() -> Info:
-    """
-    Returns a namespace with Array API namespace inspection utilities.
-
-    Returns
-    -------
-    out: Info
-        Namespace with Array API namespace inspection utilities.
-    """
-
-
-class Info:
-    def __init__(self: Info) -> None:
-        """Initialize the attributes for the inspection API class."""
-
-    @property
-    def capabilities(self: Info) -> Capabilities:
-        """
-        Dictionary of array library capabilities.
-
-        The dictionary should contain the following keys:
-
-        -   **boolean_indexing**: boolean indicating whether an array library supports boolean indexing.
-        -   **data_dependent_shapes**: boolean indicating whether an array library supports data-dependent output shapes.
-
-        Returns
-        -------
-        out: Capabilities
-            a dictionary of array library capabilities.
-        """
-
-    def default_device(self: Info) -> device:
-        """
-        Returns an object for the default device.
-
-        Returns
-        -------
-        out: device
-            an object corresponding to the default device.
-        """
-
-    def default_dtypes(
-        self: Info,
-        /,
-        *,
-        device: Optional[device] = None,
-    ) -> DefaultDataTypes:
-        """
-        Returns a dictionary containing default data types.
-
-        The dictionary should have the following keys:
-
-        -   **real floating**: default real floating-point data type.
-        -   **complex floating**: default complex floating-point data type.
-        -   **integral**: default integral data type.
-        -   **indexing**: default index data type.
-
-        Parameters
-        ----------
-        device: Optional[device]
-            device for which to return default data types. If ``device`` is ``None``, the returned data types must be the default data types for the current device. If ``device`` is not ``None``, the returned data types must be default data types specific to the specified device. Default: ``None``.
-
-            .. note::
-               Some array libraries have the concept of a device context manager, allowing library consumers to manage the current device context. When ``device`` is ``None``, libraries supporting a device context should return the default data types for the current device. For libraries without a context manager or supporting only a single device, those libraries should return the default data types for the default device.
-
-        Returns
-        -------
-        out: DefaultDataTypes
-            a dictionary containing the default data type for respective data type kinds.
-        """
-
-    def dtypes(
-        self: Info,
-        /,
-        *,
-        device: Optional[device] = None,
-        kind: Optional[Union[str, Tuple[str, ...]]] = None,
-    ) -> DataTypes:
-        """
-        Returns a dictionary of supported *Array API* data types.
-
-        .. note::
-           While specification-conforming array libraries may support additional data types which are not present in this specification, data types which are not present in this specification should not be included in the returned dictionary.
-
-        Parameters
-        ----------
-        kind: Optional[Union[str, Tuple[str, ...]]]
-            data type kind.
-
-            -   If ``kind`` is ``None``, the function must return a dictionary containing all supported Array API data types.
-
-            -   If ``kind`` is a string, the function must return a dictionary containing the data types belonging to the specified data type kind. The following data type kinds must be supported:
-
-                -   ``'bool'``: boolean data types (e.g., ``bool``).
-                -   ``'signed integer'``: signed integer data types (e.g., ``int8``, ``int16``, ``int32``, ``int64``).
-                -   ``'unsigned integer'``: unsigned integer data types (e.g., ``uint8``, ``uint16``, ``uint32``, ``uint64``).
-                -   ``'integral'``: integer data types. Shorthand for ``('signed integer', 'unsigned integer')``.
-                -   ``'real floating'``: real-valued floating-point data types (e.g., ``float32``, ``float64``).
-                -   ``'complex floating'``: complex floating-point data types (e.g., ``complex64``, ``complex128``).
-                -   ``'numeric'``: numeric data types. Shorthand for ``('integral', 'real floating', 'complex floating')``.
-
-            -   If ``kind`` is a tuple, the tuple specifies a union of data type kinds, and the function must return a dictionary containing the data types belonging to at least one of the specified data type kinds.
-
-            Default: ``None``.
-        device: Optional[device]
-            device for which to return supported data types. If ``device`` is ``None``, the returned data types must be the supported data types for the current device; otherwise, the returned data types must be supported data types specific to the specified device. Default: ``None``.
-
-            .. note::
-               Some array libraries have the concept of a device context manager, allowing library consumers to manage the current device context. When ``device`` is ``None``, libraries supporting a device context should return the supported data types for the current device. For libraries without a context manager or supporting only a single device, those libraries should return the supported data types for the default device.
-
-        Returns
-        -------
-        out: DataTypes
-            a dictionary containing supported data types.
-        """
-
-    def device(self: Info) -> List[device]:
-        """
-        Returns a list of supported devices.
-
-        Returns
-        -------
-        out: List[device]
-            a list of supported devices.
-        """
